xplora_gps_recorder/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
alembic>=1.14,<2.0
pydantic-settings>=2.6,<3.0
apscheduler>=3.10,<4.0
requests>=2.32,<3.0
python-json-logger>=2.0,<3.0
paho-mqtt>=2.1,<3.0
redis>=5.0,<6.0

